refactor: move genSuggestionOptions
