+++ packages/cspell-trie-lib/src/lib/suggestions/genSuggestionsOptions.ts
@@ -1,5 +1,5 @@
-import { WeightMap } from '.';
-import { CompoundWordsMethod } from './walker';
+import { WeightMap } from '..';
+import { CompoundWordsMethod } from '../walker';
 
 export interface GenSuggestionOptionsStrict {
     /**
@@ -18,6 +18,12 @@ export interface GenSuggestionOptionsStrict {
      * 3 is a good number. Above 5 can be very slow.
      */
     changeLimit: number;
+
+    /**
+     * Include the `+` compound character when returning results.
+     * @default false
+     */
+    includeCompoundChar?: boolean;
 }
 
 export type GenSuggestionOptions = Partial<GenSuggestionOptionsStrict>;
@@ -57,6 +63,7 @@ export const defaultGenSuggestionOptions: GenSuggestionOptionsStrict = {
     compoundMethod: CompoundWordsMethod.NONE,
     ignoreCase: true,
     changeLimit: 5,
+    includeCompoundChar: false,
 };
 
 export const defaultSuggestionOptions: SuggestionOptionsStrict = {
@@ -82,6 +89,7 @@ const keyMapOfGenSuggestionOptionsStrict: KeyMapOfGenSuggestionOptionsStrict = {
     changeLimit: 'changeLimit',
     compoundMethod: 'compoundMethod',
     ignoreCase: 'ignoreCase',
+    includeCompoundChar: 'includeCompoundChar',
 } as const;
 
 const keyMapOfSuggestionOptionsStrict: KeyMapOfSuggestionOptionsStrict = {
